core: drop extra info from trace/debug/error logging
For the trace, debug, and error levels, we were logging the filename,
line number, function name, and time with microseconds in addition to
the log message.

This is annoying because it makes messages at those levels not line up
with messages at other levels, and if the messages require
file/line/func disambiguation, then they should just include that
information in the message themselves. Also, you can get microsecond
resolution from the journal with "journalctl -o short-precise".

In addition to not being terribly useful, this also adds lots of
unnecessary strings to the binary.

So drop that extra info.