platform: implement NMPObject and NMPCache
Parts of the implementation is inside nm-linux-platform.c,
because it depends on utility functions from there.

NMPObject is a simple "object" implemenation around NMPlatformObject.
They are ref-counted and have a class-pointer. Several basic functions
like equality, hash, to-string are implemented.

NMPCache is can be used to store the NMPObject. Objects are indexed
via their primary id, but there is also multi-lookup via NMCacheId
and NMMultiIndex.