core: add NMRefString
NMRefString is a simple, refcounted, immutable string. Increasing/decreasing
the refcount does not affect const-ness.

It can be used just like a regular 'const char *' pointer. The only
difference is that you need special alloc/free functions.

It will be used for NMPlatformLink.udi field, which currently reuses the
internal value from the GUdevDevice instance. Later we need to keep our
own clone of that string, so let's make it ref-counted.
-- especially, because we create throw-away clones of link objects while
updating, and because we compare those clones to the original afterwards
(which simplifies then to a pointer equality check).