nm-dispatcher: always run pre-up and pre-down scripts immediately
pre-up and pre-down events are handled synchronously and this means
that NM has to wait for them to finish before progressing to the next
state; but the scripts for these events end up in the same dispatching
queue as other (asynchronous) events and thus may have to wait
potentially long times.

Make synchronous requests independent from the queue and execute them
immediately to minimize device activation and deactivation times. The
downside of this change is that pre-up and pre-down scripts are no
longer serialized and need to be written carefully to avoid race
conditions.

https://bugzilla.gnome.org/show_bug.cgi?id=746703