config: don't log warning in nm_config_keyfile_get_boolean()
We don't want to log a warning every time we access a keyfile value.
If we would like to warn about invalid values, we should do it once
after the configuration is loaded. And then we should not only do it
for booleans, but for other types as well.

For now, just drop this.