platform: drop nm_platform_get_error()
For NMPlatform instances we had a error reporting mechanism
to store the last error and check it via nm_platform_get_error().

Remove this. It was not used much, and it is not a great way
to report errors.

One problem is that at the point where the error happens, you don't
know whether anybody care about an error code. So, you add code to set
the error reason because somebody *might* need it (but in realitiy, almost
no caller cares).
Also, we added that to the test, testing functionality that isn't
even used in production code. But at the same time, there are almost
no users (beside testing) of this error reason, so it's almost certainly
buggy.

Then, sometimes we call into multiple platform functions that might
overwrite the error. So, every function must be cautious to preserve/set
the code for its own.

It is better to return the error reason as it is actually used. For that we already
have our common glib patterns

    (1) gboolean (*fcn) (...);
    (2) gboolean (*fcn) (..., GError **error);

In certain cases, we need more result then a #gboolean, but don't want
to bother constructing a #GError. So we should do instead:

    (3) NMPlatformError (*fcn) (...);